refactor(cymbal): carve shared symbolication kernel into core/
Move the modules both run modes share into a new src/core/ subtree:
symbolication/, langs/, frames/, error.rs, metric_consts.rs, plus new
core::ids (TeamId), core::sanitize (moved out of lib.rs), and
core::analytics (symbol-set capture split out of posthog_utils).

Crate-root re-export shims keep existing import paths resolving, so this
phase is a physical reorg with no call-site churn. posthog_utils keeps
only issue-lifecycle captures; types::operator re-exports TeamId from
core::ids. No behavior change; cargo check + clippy green.

Part of the modes/core layout refactor (phase 1 of 3).
