+++ rust/cymbal/src/core/error.rs
