+++ rust/cymbal/src/core/frames/mod.rs
