+++ rust/cymbal/src/core/ids.rs
@@ -1,0 +1,2 @@
+/// Tenant identifier threaded through symbol resolution and the pipeline.
+pub type TeamId = i32;
