+++ rust/cymbal/src/core/mod.rs
@@ -1,0 +1,13 @@
+//! Shared symbolication kernel: the symbol-resolution stack and the domain
+//! pieces both run modes depend on (see `crate::modes`). Nothing here may
+//! depend on processing-only modules (the pipeline, Kafka/Redis, rules), so
+//! this subtree can later be lifted into its own `cymbal-core` crate.
+
+pub mod analytics;
+pub mod error;
+pub mod frames;
+pub mod ids;
+pub mod langs;
+pub mod metric_consts;
+pub mod sanitize;
+pub mod symbolication;
