+++ rust/cymbal/src/posthog_utils.rs
@@ -1,11 +1,10 @@
 use posthog_rs::Event;
-use tracing::error;
 use uuid::Uuid;
+
+use crate::core::analytics::spawning_capture;
 
 const ISSUE_CREATED: &str = "error_tracking_issue_created";
 const ISSUE_REOPENED: &str = "error_tracking_issue_reopened";
-const SYMBOL_SET_SAVED: &str = "error_tracking_symbol_set_saved";
-const SYMBOL_SET_DELETED: &str = "error_tracking_symbol_set_deleted";
 
 pub fn capture_issue_created(team_id: i32, issue_id: Uuid, sentry_integration: bool) {
     let mut event = Event::new_anon(ISSUE_CREATED);
@@ -22,35 +21,4 @@ pub fn capture_issue_reopened(team_id: i32, issue_id: Uuid) {
     event.insert_prop("team_id", team_id).unwrap();
     event.insert_prop("issue_id", issue_id.to_string()).unwrap();
     spawning_capture(ISSUE_REOPENED, event);
-}
-
-pub fn capture_symbol_set_saved(team_id: i32, set_ref: &str, storage_ptr: &str, was_retry: bool) {
-    let mut event = Event::new_anon(SYMBOL_SET_SAVED);
-    event.insert_prop("team_id", team_id).unwrap();
-    event.insert_prop("set_ref", set_ref).unwrap();
-    event.insert_prop("storage_ptr", storage_ptr).unwrap();
-    event.insert_prop("was_retry", was_retry).unwrap();
-    spawning_capture(SYMBOL_SET_SAVED, event);
-}
-
-pub fn capture_symbol_set_deleted(team_id: i32, set_ref: &str, storage_ptr: Option<&str>) {
-    let mut event = Event::new_anon(SYMBOL_SET_DELETED);
-    event.insert_prop("team_id", team_id).unwrap();
-    event.insert_prop("set_ref", set_ref).unwrap();
-    if let Some(ptr) = storage_ptr {
-        event.insert_prop("storage_ptr", ptr).unwrap();
-    }
-    spawning_capture(SYMBOL_SET_DELETED, event);
-}
-
-pub fn spawning_capture(event_name: &'static str, event: Event) {
-    if posthog_rs::global_is_disabled() {
-        return;
-    }
-
-    tokio::spawn(async move {
-        if let Err(e) = posthog_rs::capture(event).await {
-            error!(event = event_name, error = ?e, "Error capturing PostHog event");
-        }
-    });
 }
